Refactored doc-sync code into a new fixture
