+++ src/integrationTest/kotlin/com/sourcegraph/cody/AllSuites.kt
@@ -3,10 +3,10 @@ package com.sourcegraph.cody
 import com.intellij.openapi.diagnostic.Logger
 import com.sourcegraph.cody.agent.CodyAgentService
 import com.sourcegraph.cody.util.CodyIntegrationTestFixture
+import java.util.concurrent.TimeUnit
 import org.junit.AfterClass
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
-import java.util.concurrent.TimeUnit
 
 /**
  * We need a single tearDown() method running after all tests are complete, so agent can be closed
@@ -21,7 +21,7 @@ import java.util.concurrent.TimeUnit
  * and define unique CODY_RECORDING_NAME.
  */
 @RunWith(Suite::class)
-@Suite.SuiteClasses(/* DocumentCodeTest::class, */DocumentSynchronizationTest::class)
+@Suite.SuiteClasses(/* DocumentCodeTest::class, */ DocumentSynchronizationTest::class)
 class AllSuites {
   companion object {
     private val logger = Logger.getInstance(AllSuites::class.java)
